add web post

+++ Assets/GameTrack/Runtime/GameTrackSDK.cs
@@ -10,7 +10,7 @@ using UnityEngine.SceneManagement;
 public class GameTrackSDK : MonoBehaviour
 {
     [DllImport("track")]
-    private static extern IntPtr GameTrack_Init(string persistentDataPath, string trackUuid, string gpuVendor);
+    private static extern IntPtr GameTrack_Init(string persistentDataPath, string track_uuid, string base_info);
     
     [DllImport("track")]
     private static extern void GameTrack_Update(float unscaledDeltaTime, int targetFrameRate);
@@ -30,7 +30,7 @@ public class GameTrackSDK : MonoBehaviour
     // Init GamePerf SDK
     private void Start()
     {
-#if UNITY_ANDROID && !UNITY_EDITOR
+#if UNITY_ANDROID //&& !UNITY_EDITOR
         // UUID
         var localUUID = PlayerPrefs.GetString("track_uuid");
         if (string.IsNullOrEmpty(localUUID))
@@ -41,15 +41,17 @@ public class GameTrackSDK : MonoBehaviour
         
         // BaseInfo
         StringBuilder baseInfo = new StringBuilder();
-        baseInfo.AppendFormat("{0},{1},{2},{3},{4},{5},{6}", Application.identifier, SystemInfo.operatingSystem, SystemInfo.deviceModel, SystemInfo.deviceName, SystemInfo.graphicsDeviceVendor, SystemInfo.deviceName, SystemInfo.graphicsDeviceVersion);
+        baseInfo.AppendFormat("{0},{1},{2},{3},{4},{5},{6}", Application.identifier.GetHashCode(), Application.identifier, SystemInfo.operatingSystem, SystemInfo.deviceModel, SystemInfo.deviceName, SystemInfo.graphicsDeviceVendor, SystemInfo.deviceName, SystemInfo.graphicsDeviceVersion);
         
         // Init
         var _logFile = GameTrack_Init(Application.persistentDataPath, localUUID, baseInfo.ToString());
         string logFile = Marshal.PtrToStringAnsi(_logFile);
         // Upload Last Files
-        StartCoroutine(UploadData(logFile));
+        // StartCoroutine(UploadData(logFile));
         // send to minio
         StartCoroutine(MinioUpdateFile(logFile));
+        // send to web
+        StartCoroutine(WebPostUpdateFile(logFile));
 
         // Track Scene
         SceneManager.sceneLoaded += SceneLoadedTrack;
@@ -70,8 +72,6 @@ public class GameTrackSDK : MonoBehaviour
     // Track User Click 
     public void UserClickTrack(string eventName)
     {
-        //Demo: GameTrack_Event("Button1"); 
-        //Todo: @lixiaofeng
         GameTrack_Event(eventName); 
     }
 
@@ -126,6 +126,39 @@ public class GameTrackSDK : MonoBehaviour
             {
                 string text = www.downloadHandler.text;
                 Debug.Log("minio upload succeed:" + text);
+                File.Delete(file.FullName);
+            }
+        }
+    }
+
+    IEnumerator WebPostUpdateFile(string currentFile)
+    {
+        DirectoryInfo directory = new DirectoryInfo(Application.persistentDataPath + "/track_data");
+        foreach (FileInfo file in directory.GetFiles())
+        {
+            if (file.FullName.Equals(currentFile))
+                continue;
+            if (file.Length == 0)
+            {
+                File.Delete(file.FullName);
+                Debug.LogFormat("File size is 0 delete: {0}", file);
+                continue;
+            }
+            if (!File.Exists(file.FullName))
+            {
+                Debug.LogFormat("File does not exist: {0}", file);
+                continue;
+            }
+            UnityWebRequest www = HTTPUtils.CreateUploadFileRequest(file.FullName, file.Name);
+            yield return www.SendWebRequest();
+            if (www.result != UnityWebRequest.Result.Success)
+            {
+                Debug.LogFormat("web post upload file: {0} error: {1} / {2}", file, www.error, www.result);
+            }
+            else
+            {
+                string text = www.downloadHandler.text;
+                Debug.Log("web post upload succeed:" + text);
                 File.Delete(file.FullName);
             }
         }
